chore: split-up helper-functions & added inline-comments to them (#374)
