+++ plugins/module_utils/helper/match_helper_pytest.py
@@ -1,0 +1,11 @@
+
+def test_sanitize_module_args():
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.match import _sanitize_module_args
+
+
+def test_get_matching():
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.match import get_matching
+
+
+def test_get_multiple_matching():
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.match import get_multiple_matching
