+++ plugins/module_utils/main/dhcp_subnet_v4.py
@@ -2,7 +2,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
-from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
     get_selected_list, simplify_translate
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.module import BaseModule
