+++ plugins/module_utils/main/ipsec_cert.py
@@ -4,7 +4,9 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
-    get_selected, is_unset
+    is_unset
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
+    get_selected
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.module import BaseModule
