+++ plugins/module_utils/main/openvpn_client.py
@@ -2,7 +2,7 @@ from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
-from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
     get_key_by_value_from_selection, get_key_by_value_end_from_selection
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.validate import \
     is_unset
