+++ plugins/module_utils/main/privilege.py
@@ -1,6 +1,6 @@
 from ansible.module_utils.basic import AnsibleModule
 
-from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
     get_key_by_value_from_selection
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
