+++ plugins/module_utils/main/route.py
@@ -2,7 +2,7 @@ from ipaddress import ip_network
 
 from ansible.module_utils.basic import AnsibleModule
 
-from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
     simplify_translate
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
