+++ plugins/module_utils/main/user.py
@@ -1,7 +1,9 @@
 from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.main import \
-    is_unset, get_key_by_value_from_selection
+    is_unset
+from ansible_collections.oxlorg.opnsense.plugins.module_utils.helper.translate import \
+    get_key_by_value_from_selection
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.api import \
     Session
 from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.module import BaseModule
